[Asset] moved the Asset classes from the Templating Component to a new one
