+++ src/Symfony/Component/Asset/CHANGELOG.md
@@ -1,0 +1,7 @@
+CHANGELOG
+=========
+
+2.7.0
+-----
+
+ * added the component
