+++ src/Symfony/Component/Asset/README.md
@@ -1,0 +1,13 @@
+Asset Component
+===============
+
+Asset provides classes to manage asset URLs.
+
+Resources
+---------
+
+You can run the unit tests with the following command:
+
+    $ cd path/to/Symfony/Component/Asset/
+    $ composer.phar install
+    $ phpunit
