+++ src/Symfony/Component/Templating/Asset/PackageInterface.php
@@ -11,27 +11,15 @@
 
 namespace Symfony\Component\Templating\Asset;
 
+use Symfony\Component\Asset\PackageInterface as AssetPackageInterface;
+
 /**
  * Asset package interface.
  *
  * @author Kris Wallsmith <kris@symfony.com>
+ *
+ * @deprecated since 2.7, to be removed in 3.0. Use the Asset Component instead.
  */
-interface PackageInterface
+interface PackageInterface extends AssetPackageInterface
 {
-    /**
-     * Returns the asset package version.
-     *
-     * @return string The version string
-     */
-    public function getVersion();
-
-    /**
-     * Returns an absolute or root-relative public path.
-     *
-     * @param string           $path    A path
-     * @param string|bool|null $version A specific version for the path
-     *
-     * @return string The public path
-     */
-    public function getUrl($path, $version = null);
 }
